- Updated imports in `qq_chat_agent.rs` and `qq_chat_agent_core.rs` to use `zihuan_graph_engine` for brain tool definitions. - Enhanced message batching logic in `qq_chat_agent.rs` to handle replies more effectively. - Introduced new nodes: `AgentImageDbRefNode`, `AgentMySqlRefNode`, `AgentRustfsRefNode`, and `AgentTavilyRefNode` for managing connections to various resources. - Implemented `BrainNode` for executing multi-turn tool invocation reasoning using LLM and built-in tools. - Added `tool_subgraph.rs` to manage tool execution and validation. - Updated `resource_resolver.rs` to reflect changes in LLM node imports. - Improved error handling and validation across nodes and tool definitions.
