+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["node_macros", "general_wheel_cpp", "zihuan_core", "zihuan_agent", "zihuan_graph_engine", "zihuan_llm", "zihuan_core_nodes", "ims_bot_adapter", "storage_handler", "zihuan_service", "zihuan_graph_cli"]
+members = ["node_macros", "general_wheel_cpp", "zihuan_core", "zihuan_agent", "zihuan_graph_engine", "zihuan_llm", "ims_bot_adapter", "storage_handler", "zihuan_service", "zihuan_graph_cli"]
 
 [patch.crates-io]
 candle-transformers = { path = "third_party/candle-transformers" }
@@ -13,8 +13,8 @@ authors = ["FredYakumo"]
 
 [features]
 default = []
-candle-cuda = ["zihuan_llm/candle-cuda", "zihuan_core_nodes/candle-cuda"]
-candle-metal = ["zihuan_llm/candle-metal", "zihuan_core_nodes/candle-metal"]
+candle-cuda = ["zihuan_llm/candle-cuda"]
+candle-metal = ["zihuan_llm/candle-metal"]
 
 [dependencies]
 tokio = { version = "1", features = ["full"] }
@@ -43,7 +43,6 @@ zihuan_core = { path = "zihuan_core" }
 zihuan_agent = { path = "zihuan_agent" }
 zihuan_graph_engine = { path = "zihuan_graph_engine" }
 zihuan_llm = { path = "zihuan_llm" }
-zihuan_core_nodes = { path = "zihuan_core_nodes" }
 ims_bot_adapter = { path = "ims_bot_adapter" }
 storage_handler = { path = "storage_handler" }
 zihuan_service = { path = "zihuan_service" }
