+++ document/node/node-development.zh-CN.md
@@ -37,8 +37,10 @@
 |---|---|
 | 通用工具 / 变换 | `zihuan_graph_engine/src/util/` |
 | 数据库 / 存储辅助 | `zihuan_graph_engine/src/` 下对应模块 |
+| 存储 / 连接 / 搜索 | `storage_handler/src/` |
 | Bot 相关同步辅助 | `ims_bot_adapter/src/` |
-| LLM / Brain / RAG | `zihuan_llm/src/` |
+| LLM / 向量化 / Agent 配置 | `zihuan_llm/src/nodes/` |
+| Brain / Agent | `zihuan_service/src/nodes/` |
 
 ### 3. 提供构造函数
 
@@ -81,7 +83,7 @@ pub fn new(id: String, name: String) -> Self
 ### 7. 注册节点
 
 - `zihuan_graph_engine` 中的节点 → `zihuan_graph_engine/src/registry.rs`
-- `ims_bot_adapter` / `zihuan_llm` 中的节点 → `src/init_registry.rs`
+- `storage_handler` / `ims_bot_adapter` / `zihuan_llm` / `zihuan_service` 中的节点 → 在所属 crate 的 `init_node_registry()` 中调用 `register_node!`，该函数由 `src/init_registry.rs` 统一引入
 
 ### 8. 验证行为
