+++ storage_handler/src/weaviate_image_search_node.rs
