--- zihuan_core_nodes/src/brain/brain_tool.rs
@@ -1,7 +1,0 @@
-pub use zihuan_graph_engine::brain_tool_spec::{
-    brain_shared_inputs_from_value, brain_tool_input_signature, fixed_tool_runtime_inputs,
-    is_tool_subgraph_owner, normalized_tool_outputs_for_owner, BrainToolDefinition, ToolParamDef,
-    BRAIN_SHARED_INPUTS_PORT, BRAIN_TOOLS_CONFIG_PORT, BRAIN_TOOL_FIXED_CONTENT_INPUT,
-    QQ_AGENT_TOOL_FIXED_BOT_ADAPTER_INPUT, QQ_AGENT_TOOL_FIXED_MESSAGE_EVENT_INPUT,
-    QQ_AGENT_TOOL_OUTPUT_NAME, QQ_AGENT_TOOL_OWNER_TYPE, QQ_AGENT_TOOL_OWNER_TYPE_LEGACY,
-};
