--- zihuan_core_nodes/src/brain/mod.rs
@@ -1,3 +1,0 @@
-pub mod brain_node;
-pub mod brain_tool;
-pub mod tool_subgraph;
