--- zihuan_core_nodes/src/linalg/mod.rs
@@ -1,5 +1,0 @@
-pub mod embedding_api;
-pub mod text_embedding;
-pub mod batch_text_embedding;
-pub mod vector_cosine_similarity;
-pub mod top_k_similarity;
