--- zihuan_core_nodes/src/nn/mod.rs
@@ -1,1 +1,0 @@
-pub mod local_candle_embedding_node;
