--- zihuan_core_nodes/src/rag/mod.rs
@@ -1,3 +1,0 @@
-pub mod tavily_provider;
-pub mod tavily_search;
-pub mod weaviate_image_search;
