+++ zihuan_graph_cli/Cargo.toml
@@ -9,7 +9,7 @@ tokio = { version = "1", features = ["full"] }
 zihuan_core = { path = "../zihuan_core" }
 zihuan_graph_engine = { path = "../zihuan_graph_engine" }
 zihuan_llm = { path = "../zihuan_llm" }
-zihuan_core_nodes = { path = "../zihuan_core_nodes" }
+zihuan_service = { path = "../zihuan_service" }
 ims_bot_adapter = { path = "../ims_bot_adapter" }
 storage_handler = { path = "../storage_handler" }
