+++ zihuan_llm/src/nodes/text_embedding_node.rs
