+++ zihuan_service/src/nodes/agent_image_db_ref.rs
@@ -1,10 +1,9 @@
 use std::collections::HashMap;
 
+use storage_handler::{load_connections, resource_resolver, WeaviateCollectionSchema};
 use zihuan_core::agent_config::current_qq_chat_agent_config;
 use zihuan_core::error::Result;
 use zihuan_graph_engine::{node_output, DataValue, DataType, Node, Port};
-
-use crate::{load_connections, resource_resolver, WeaviateCollectionSchema};
 
 pub struct AgentImageDbRefNode {
     id: String,
