+++ zihuan_service/src/resource_resolver.rs
@@ -5,7 +5,7 @@ use zihuan_core::error::{Error, Result};
 use zihuan_core::llm::embedding_base::EmbeddingBase;
 use zihuan_core::llm::llm_base::LLMBase;
 use zihuan_llm::linalg::embedding_api::EmbeddingAPI;
-use zihuan_core_nodes::llm::llm_node::build_llm;
+use zihuan_llm::nodes::llm_node::build_llm;
 use zihuan_llm::nn::queued_embedding_model::QueuedEmbeddingModel;
 use zihuan_core::agent_config::EmbeddingServiceConfig;
 use zihuan_llm::system_config::{LlmRefConfig, LlmServiceConfig, ModelRefSpec};
